+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog] and this project adheres to
 their local inner macros. Not having this prevented them from working when
 `use`d or called directly via `static_assertions::macro!(...)`.
 
+### Removed
+- Unused `_assert_obj_safe!` from pre-1.0
+
 ## [1.0.0]: - 2019-10-02
 ### Added
 - `assert_eq_align!` macro